fix: Se ha abstraído el componente de feature fuera de la vista de la pantalla principal (index)
