+++ frontend/app/(drawer)/index.tsx
@@ -9,33 +9,9 @@ import { Card, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Text } from "@/components/ui/text";
 import { Textarea } from "@/components/ui/textarea";
+import { Feature } from "@/components/feature";
 
 import { API_URL, getGlobalJwtToken } from "@/constants/api";
-
-interface FeatureProps {
-  Icon: LucideIcon;
-  title: string;
-  desc: string;
-}
-
-function Feature({ Icon, title, desc }: FeatureProps) {
-  return (
-    <Card className="w-full md:w-[48%]">
-      <CardContent>
-        <View className="flex-row items-center gap-4">
-          <View className="w-14 h-14 rounded-xl items-center justify-center bg-blue-100 dark:bg-blue-900/30">
-            <Icon size={26} color="#5c90cf" />
-          </View>
-
-          <View className="flex-1">
-            <Text className="font-semibold">{title}</Text>
-            <Text className="text-muted-foreground">{desc}</Text>
-          </View>
-        </View>
-      </CardContent>
-    </Card>
-  );
-}
 
 export default function HomeScreen() {
   const navigation = useNavigation<DrawerNavigationProp<ParamListBase>>();
